Correct bbox calculation
The original code mixed the name of the model matrix in MVP
transformation and the matrix of the class 'Model'. This caused the
confusion when doing the transformation. Thus this commit clarify the
name of these matrix and function at first.

The main problem of the bbox now is that Skinned models were computing
their bounding boxes with the mesh's node transform applied in CPU code.

However, the vertex shader already multiplies each vertex by
'jointMatrices', which include the node transformations, before applying
the model matrix.

Because the CPU code also multiplied each mesh's local bounding box by
the node matrix, the transformation was effectively applied twice,
producing oversized and offset bounding boxes.

This commit add several utility for bbox, it would detected that if the
model contained animation, then called the corresponding bbox
calculation.