Implement character movement
This commit adds a new ImGUI checkbox that enables a movement mode for
the character. In this mode, the camera follows the character, and the
character moves based on the camera's facing direction and keyboard
input.

Additionally, this commit binds the name in the animation state to the
currently selected model in the scene entities, so the model selection
dropdown in the animation control window has been removed.